refactor: extract build-info to test folder
