+++ packages/core/src/lib/file-info/unassigned-file-info.ts
@@ -1,9 +1,8 @@
-import { FsPath, toFsPath } from './fs-path';
+import { FsPath } from './fs-path';
 import throwIfNull from '../util/throw-if-null';
-import getFs from '../fs/getFs';
 
 /**
- * Class representing a TypeScript file with its dependencies.
+ * Represents a TypeScript file with its dependencies.
  * If an import cannot be resolved, it doesn't throw an error
  * but is added to unresolvableImports.
  *
@@ -42,52 +41,3 @@ export default class UnassignedFileInfo {
     );
   }
 }
-
-type NestedArray = (string | NestedArray)[];
-
-const createPath = (path: string, parentPath: string) => {
-  let currentPath = path;
-  if (path.startsWith('./')) {
-    const parentPaths = parentPath.split('/');
-    parentPaths.pop();
-    const normalisedPaths = path.split('/').slice(1);
-
-    currentPath = [...parentPaths, ...normalisedPaths].join('/');
-  }
-
-  return currentPath;
-};
-
-/**
- * utility function generate a FileInfo Tree for testing purposes
- * @param path
- * @param imports
- */
-export const buildFileInfo = (
-  path: string,
-  imports: NestedArray = [],
-): UnassignedFileInfo => {
-  const fs = getFs();
-  const children: UnassignedFileInfo[] = imports.map((entry) => {
-    if (
-      Array.isArray(entry) &&
-      entry.length === 2 &&
-      typeof entry[0] === 'string'
-    ) {
-      const [childPath, childImports] = entry;
-      const currentPath = createPath(childPath, path);
-      return buildFileInfo(currentPath, childImports as NestedArray[]);
-    } else if (typeof entry === 'string') {
-      return buildFileInfo(createPath(entry, path), []);
-    } else {
-      throw new Error(`import not in right structure`);
-    }
-  });
-
-  fs.writeFile(path, '');
-  const fileInfo = new UnassignedFileInfo(toFsPath(path));
-  for (const child of children) {
-    fileInfo.addImport(child, fs.relativeTo(fileInfo.path, child.path));
-  }
-  return fileInfo;
-};
